Move utils.py to tests folder
