+++ tests/scenario/test_gazebo_simulator.py
@@ -3,8 +3,8 @@
 # GNU Lesser General Public License v2.1 or any later version.
 
 import pytest
-from . import utils
-from .utils import gazebo_fixture as gazebo
+from .. import utils
+from ..utils import gazebo_fixture as gazebo
 from gym_ignition import scenario_bindings as bindings
 
 # Set the verbosity
